将banner移入精选页
